requirements.txt
fastapi
uvicorn
librosa
vosk
python-dotenv
python-multipart